requirements.txt
annotated-types==0.7.0
anyio==4.4.0
attrs==23.2.0
beautifulsoup4==4.12.3  # only used by persona generation; kept
certifi==2024.7.4
click==8.1.7
cloudpickle==3.0.0
distro==1.9.0
dominate==2.9.1  # legacy; safe to remove but kept if referenced in history
farama-notifications==0.0.4  # legacy; safe to remove
gymnasium==0.29.1  # legacy; safe to remove
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
ipywidgets==8.1.5
json-fix==1.0.0
jupyterlab-widgets==3.0.13
regex==2024.11.6
numpy==2.0.1  # used by memory
openai==1.38.0
outcome==1.3.0.post0
pandas==2.2.2
pydantic==2.8.2
pydantic-core==2.20.1
pysocks==1.7.1
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.2
selenium==4.23.1  # legacy; remove
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
tqdm==4.66.4  # used by agent and persona
trio==0.26.0
trio-websocket==0.11.1
tzdata==2024.1
urllib3==2.2.2
websocket-client==1.8.0
widgetsnbextension==4.0.13
wsproto==1.2.0
ipython==8.18.1
matplotlib-inline==0.1.7
# AgentQL Dependencies
agentql==1.0.11
playwright==1.47.0
playwright-stealth==1.0.6 